[Feat] #10 - Divide local component
