+++ app/src/main/java/kr/yhs/traffic/ui/pages/StationListPage.kt
@@ -1,21 +1,10 @@
 package kr.yhs.traffic.ui.pages
 
 import android.location.Location
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.wear.compose.foundation.lazy.ScalingLazyColumn
 import androidx.wear.compose.foundation.lazy.ScalingLazyListState
@@ -25,9 +14,10 @@ import androidx.wear.compose.material.*
 import com.google.android.horologist.annotations.ExperimentalHorologistApi
 import com.google.android.horologist.compose.rotaryinput.rotaryWithSnap
 import com.google.android.horologist.compose.rotaryinput.toRotaryScrollAdapter
-import kr.yhs.traffic.R
 import kr.yhs.traffic.models.StationInfo
 import kr.yhs.traffic.ui.components.LoadingProgressIndicator
+import kr.yhs.traffic.ui.components.StationEmpty
+import kr.yhs.traffic.ui.components.StationShortInfo
 import kr.yhs.traffic.ui.components.WearScaffold
 import kotlin.math.atan2
 import kotlin.math.roundToInt
@@ -116,88 +106,4 @@ fun StationListPage(
             }
         }
     }
-}
-
-
-@Composable
-fun StationEmpty() {
-    val context = LocalContext.current
-    return Column(
-        modifier = Modifier.fillMaxSize(), horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.ic_baseline_search_off),
-            contentDescription = "Not Found",
-            modifier = Modifier.height(30.dp)
-        )
-        Text(
-            text = context.getString(R.string.result_not_found),
-            style = MaterialTheme.typography.body1
-        )
-    }
-}
-
-
-@Composable
-fun StationShortInfo(
-    stationName: String,
-    stationId: String,
-    distance: Int = -1,
-    direction: Int = -1,
-    call: () -> Unit
-) {
-    Chip(
-        modifier = Modifier
-            .height(54.dp)
-            .fillMaxWidth()
-            .padding(top = 2.dp, bottom = 2.dp),
-        colors = ChipDefaults.chipColors(
-            backgroundColor = Color.White, contentColor = Color.Black
-        ),
-        label = {
-            Text(
-                text = stationName,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                style = MaterialTheme.typography.body1
-            )
-        },
-        secondaryLabel = {
-            var secondaryText = stationId
-            if (distance in 1..500) secondaryText += " | ${distance}m"
-            Text(
-                text = secondaryText, maxLines = 1, style = MaterialTheme.typography.body2
-            )
-        },
-        icon = {
-            Box {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_baseline_directions_bus),
-                    contentDescription = "bus icon",
-                    modifier = Modifier
-                        .size(24.dp)
-                        .wrapContentSize(align = Alignment.Center)
-                        .align(Alignment.Center)
-                )
-                if (distance in 1..500) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.ic_baseline_arrow),
-                        contentDescription = "Rotation",
-                        modifier = Modifier
-                            .size(14.dp)
-                            .wrapContentSize(align = Alignment.BottomEnd)
-                            .align(Alignment.BottomEnd)
-                            .rotate(direction.toFloat())
-                            .background(Color.White)
-                            .border(
-                                width = 1.dp, brush = Brush.radialGradient(
-                                    colors = listOf(Color.Black, Color.LightGray)
-                                ), shape = CircleShape
-                            )
-                    )
-                }
-            }
-        },
-        onClick = call
-    )
 }
